requirements.txt
pandas
numpy
plotly
streamlit
biopython
python-dotenv
matplotlib
wordcloud